requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
pytesseract==0.3.9
Pillow==10.3.0
opencv-python-headless
openai
requests